[roll] Update 3p packages
googleapis     0.50.2 --> 0.50.4

+++ googleapis/lib/admin/directory_v1.dart
@@ -1,5 +1,7 @@
 // This is a generated file (see the discoveryapis_generator project).
 
+// ignore_for_file: unnecessary_cast
+
 library googleapis.admin.directory_v1;
 
 import 'dart:core' as core;
@@ -6956,10 +6958,10 @@ class BuildingCoordinates {
 
   BuildingCoordinates.fromJson(core.Map _json) {
     if (_json.containsKey("latitude")) {
-      latitude = _json["latitude"];
+      latitude = _json["latitude"].toDouble();
     }
     if (_json.containsKey("longitude")) {
-      longitude = _json["longitude"];
+      longitude = _json["longitude"].toDouble();
     }
   }
 
@@ -10090,10 +10092,10 @@ class SchemaFieldSpecNumericIndexingSpec {
 
   SchemaFieldSpecNumericIndexingSpec.fromJson(core.Map _json) {
     if (_json.containsKey("maxValue")) {
-      maxValue = _json["maxValue"];
+      maxValue = _json["maxValue"].toDouble();
     }
     if (_json.containsKey("minValue")) {
-      minValue = _json["minValue"];
+      minValue = _json["minValue"].toDouble();
     }
   }
