+++ googleapis/lib/dfareporting/v3_0.dart
@@ -1,5 +1,7 @@
 // This is a generated file (see the discoveryapis_generator project).
 
+// ignore_for_file: unnecessary_cast
+
 library googleapis.dfareporting.v3_0;
 
 import 'dart:core' as core;
@@ -17281,7 +17283,7 @@ class Conversion {
       timestampMicros = _json["timestampMicros"];
     }
     if (_json.containsKey("value")) {
-      value = _json["value"];
+      value = _json["value"].toDouble();
     }
   }
 
@@ -18319,7 +18321,7 @@ class Creative {
       videoDescription = _json["videoDescription"];
     }
     if (_json.containsKey("videoDuration")) {
-      videoDuration = _json["videoDuration"];
+      videoDuration = _json["videoDuration"].toDouble();
     }
   }
 
@@ -18929,7 +18931,7 @@ class CreativeAsset {
       pushdown = _json["pushdown"];
     }
     if (_json.containsKey("pushdownDuration")) {
-      pushdownDuration = _json["pushdownDuration"];
+      pushdownDuration = _json["pushdownDuration"].toDouble();
     }
     if (_json.containsKey("role")) {
       role = _json["role"];
@@ -18953,7 +18955,7 @@ class CreativeAsset {
       verticallyLocked = _json["verticallyLocked"];
     }
     if (_json.containsKey("videoDuration")) {
-      videoDuration = _json["videoDuration"];
+      videoDuration = _json["videoDuration"].toDouble();
     }
     if (_json.containsKey("windowMode")) {
       windowMode = _json["windowMode"];
