+++ googleapis/lib/sheets/v4.dart
@@ -1,5 +1,7 @@
 // This is a generated file (see the discoveryapis_generator project).
 
+// ignore_for_file: unnecessary_cast
+
 library googleapis.sheets.v4;
 
 import 'dart:core' as core;
@@ -3504,7 +3506,7 @@ class BubbleChartSpec {
       bubbleMinRadiusSize = _json["bubbleMinRadiusSize"];
     }
     if (_json.containsKey("bubbleOpacity")) {
-      bubbleOpacity = _json["bubbleOpacity"];
+      bubbleOpacity = _json["bubbleOpacity"].toDouble();
     }
     if (_json.containsKey("bubbleSizes")) {
       bubbleSizes = new ChartData.fromJson(_json["bubbleSizes"]);
@@ -4503,16 +4505,16 @@ class Color {
 
   Color.fromJson(core.Map _json) {
     if (_json.containsKey("alpha")) {
-      alpha = _json["alpha"];
+      alpha = _json["alpha"].toDouble();
     }
     if (_json.containsKey("blue")) {
-      blue = _json["blue"];
+      blue = _json["blue"].toDouble();
     }
     if (_json.containsKey("green")) {
-      green = _json["green"];
+      green = _json["green"].toDouble();
     }
     if (_json.containsKey("red")) {
-      red = _json["red"];
+      red = _json["red"].toDouble();
     }
   }
 
@@ -6003,7 +6005,7 @@ class ExtendedValue {
       formulaValue = _json["formulaValue"];
     }
     if (_json.containsKey("numberValue")) {
-      numberValue = _json["numberValue"];
+      numberValue = _json["numberValue"].toDouble();
     }
     if (_json.containsKey("stringValue")) {
       stringValue = _json["stringValue"];
@@ -6706,13 +6708,13 @@ class HistogramChartSpec {
 
   HistogramChartSpec.fromJson(core.Map _json) {
     if (_json.containsKey("bucketSize")) {
-      bucketSize = _json["bucketSize"];
+      bucketSize = _json["bucketSize"].toDouble();
     }
     if (_json.containsKey("legendPosition")) {
       legendPosition = _json["legendPosition"];
     }
     if (_json.containsKey("outlierPercentile")) {
-      outlierPercentile = _json["outlierPercentile"];
+      outlierPercentile = _json["outlierPercentile"].toDouble();
     }
     if (_json.containsKey("series")) {
       series = (_json["series"] as core.List)
@@ -6796,13 +6798,13 @@ class HistogramRule {
 
   HistogramRule.fromJson(core.Map _json) {
     if (_json.containsKey("end")) {
-      end = _json["end"];
+      end = _json["end"].toDouble();
     }
     if (_json.containsKey("interval")) {
-      interval = _json["interval"];
+      interval = _json["interval"].toDouble();
     }
     if (_json.containsKey("start")) {
-      start = _json["start"];
+      start = _json["start"].toDouble();
     }
   }
 
@@ -7015,7 +7017,7 @@ class IterativeCalculationSettings {
 
   IterativeCalculationSettings.fromJson(core.Map _json) {
     if (_json.containsKey("convergenceThreshold")) {
-      convergenceThreshold = _json["convergenceThreshold"];
+      convergenceThreshold = _json["convergenceThreshold"].toDouble();
     }
     if (_json.containsKey("maxIterations")) {
       maxIterations = _json["maxIterations"];
@@ -7719,7 +7721,7 @@ class PieChartSpec {
       legendPosition = _json["legendPosition"];
     }
     if (_json.containsKey("pieHole")) {
-      pieHole = _json["pieHole"];
+      pieHole = _json["pieHole"].toDouble();
     }
     if (_json.containsKey("series")) {
       series = new ChartData.fromJson(_json["series"]);
