+++ googleapis/lib/testing/v1.dart
@@ -1,5 +1,7 @@
 // This is a generated file (see the discoveryapis_generator project).
 
+// ignore_for_file: unnecessary_cast
+
 library googleapis.testing.v1;
 
 import 'dart:core' as core;
@@ -1460,7 +1462,7 @@ class Distribution {
 
   Distribution.fromJson(core.Map _json) {
     if (_json.containsKey("marketShare")) {
-      marketShare = _json["marketShare"];
+      marketShare = _json["marketShare"].toDouble();
     }
     if (_json.containsKey("measurementTime")) {
       measurementTime = _json["measurementTime"];
@@ -2887,19 +2889,19 @@ class TrafficRule {
 
   TrafficRule.fromJson(core.Map _json) {
     if (_json.containsKey("bandwidth")) {
-      bandwidth = _json["bandwidth"];
+      bandwidth = _json["bandwidth"].toDouble();
     }
     if (_json.containsKey("burst")) {
-      burst = _json["burst"];
+      burst = _json["burst"].toDouble();
     }
     if (_json.containsKey("delay")) {
       delay = _json["delay"];
     }
     if (_json.containsKey("packetDuplicationRatio")) {
-      packetDuplicationRatio = _json["packetDuplicationRatio"];
+      packetDuplicationRatio = _json["packetDuplicationRatio"].toDouble();
     }
     if (_json.containsKey("packetLossRatio")) {
-      packetLossRatio = _json["packetLossRatio"];
+      packetLossRatio = _json["packetLossRatio"].toDouble();
     }
   }
